backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.30,<1
langchain>=0.3,<1
langchain-openai>=0.2,<1
langchain-community>=0.3,<1
langchain-text-splitters>=0.3,<1
python-dotenv>=1,<2
websockets>=12,<16
openai>=1.30,<2
python-multipart>=0.0.9,<1
numpy>=1.26,<3
faiss-cpu>=1.8,<2
sentence-transformers>=3,<4
